requirements.txt
fastapi
uvicorn
chainlit
httpx
msal
python-dotenv
azure-identity
azure-storage-blob
starlette 
itsdangerous
